Added Backpropagation, Deck Manager class, And MCTS seems to work and competes fairly well with other AI.

+++ core-extensions/SabberStoneCoreAi/src/Agent/AlvaroMCTS/Node.cs
@@ -7,7 +7,7 @@ namespace SabberStoneCoreAi.src.Agent.AlvaroMCTS
 {
 	class Node
 	{
-		public int totalValue	{ get; set; } // de que gana el jugador 1.
+		public float totalValue	{ get; set; } // de que gana el jugador 1.
 		public int timesVisited { get; set; }
 		public Node parent		{ get; set; }
 		public PlayerTask task	{ get; set; }
